backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pandas==2.2.3
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
langchain==0.1.13
langchain-google-genai==1.0.3
langchain-community==0.0.29
